pyproject.toml
[project]
name = "qqautoreply"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "keyboard>=0.13.5",
    "logging>=0.4.9.6",
    "mss>=10.1.0",
    "numpy>=2.3.5",
    "openai>=2.12.0",
    "opencv-python>=4.11.0.86",
    "paddleocr>=3.3.2",
    "paddlepaddle-gpu>=3.2.2",
    "pyautogui>=0.9.54",
    "pyperclip>=1.11.0",
    "pypubsub>=4.0.3",
    "python-dotenv>=1.2.1",
    "scikit-image>=0.25.2",
]

[tool.uv]
environments = ["sys_platform == 'win32'"]

[[tool.uv.index]]
name = "paddlepaddle"
url = "https://www.paddlepaddle.org.cn/packages/stable/cu129/"
explicit = true

[tool.uv.sources]
paddlepaddle-gpu = { index = "paddlepaddle" }
